Refactor AddWordForm to replace AddWordForm2 and streamline audio handling
