+++ src/app/addWord/page.tsx
@@ -1,6 +1,6 @@
 import { getActiveUserSession } from "@/lib/auth";
 import styles from "./page.addWord.module.css";
-import AddWordForm2 from "@/components/AddWordForm2";
+import AddWordForm from "@/components/AddWordForm";
 
 export default async function AddWord() {
     await getActiveUserSession();
@@ -9,7 +9,7 @@ export default async function AddWord() {
         <main>
             <div className={styles.Container}>
                 <h2 className={styles.Title}>Lägg till nytt Dialektalt ord</h2>
-                <AddWordForm2 />
+                <AddWordForm />
             </div>
         </main>
     );
