+++ src/components/Audio.ts
@@ -1,116 +1,81 @@
-import { useState, useRef } from "react";
+import { useRef, useState } from "react";
+import { UseFormSetValue, FieldValues, Path, PathValue } from "react-hook-form";
 
-export function useAudio() {
-    const [audioFile, setAudioFile] = useState<string | null>(null);
+function useAudio2() {
     const [isRecording, setIsRecording] = useState(false);
     const mediaRecorderRef = useRef<MediaRecorder | null>(null);
-    const recordedChunksRef = useRef<Blob[]>([]);
-
-    // Convert file/blob to base64
-    function fileToBase64(file: File | Blob): Promise<string> {
-        return new Promise((resolve, reject) => {
-            if (!file) {
-                reject(new Error("File is undefined"));
-                return;
-            }
-            const reader = new FileReader();
-            reader.onload = () => {
-                if (typeof reader.result === "string") {
-                    resolve(reader.result);
-                } else {
-                    reject(new Error("Failed to convert file to base64"));
-                }
-            };
-            reader.onerror = reject;
-            reader.readAsDataURL(file);
-        });
-    }
-
-    function base64ToFileList(): FileList {
-        const file = new File(recordedChunksRef.current, "recording.mp3", { type: "audio/mp3" });
-        
-        // Skapa en FileList med det nya filobjektet
-        const dataTransfer = new DataTransfer();
-        dataTransfer.items.add(file);
-        
-        return dataTransfer.files;
-    }
+    const [audio, setAudio] = useState<HTMLAudioElement | null>(null);
+    const [recordingSoundFile, setRecordingSoundFile] = useState<File | null>(null);
 
     async function startRecording() {
-        recordedChunksRef.current = [];
-        try {
-            const stream = await navigator.mediaDevices.getUserMedia({
-                audio: true,
+        const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
+        const mediaRecorder = new MediaRecorder(stream, { mimeType: "audio/webm" });
+        mediaRecorderRef.current = mediaRecorder;
+
+        mediaRecorder.start();
+
+        setIsRecording(true);
+    }
+
+    async function stopRecording<T extends FieldValues>({
+        setValue,
+        fieldName,
+    }: {
+        setValue: UseFormSetValue<T>;
+        fieldName: Path<T>;
+    }) {
+        if (!mediaRecorderRef.current) {
+            return;
+        }
+
+        // Stoppa inspelningen, vilket kommer att trigga ondataavailable-eventet där vi hanterar den inspelade ljudfilen
+        mediaRecorderRef.current.stop();
+
+        // När data är tillgänglig, skapa en fil och spara den i state
+        mediaRecorderRef.current.ondataavailable = (event: BlobEvent) => {
+            const audioBlob = event.data;
+
+            const audioFile = new File([audioBlob], "recording.webm", {
+                type: "audio/webm",
             });
-            const mediaRecorder = new MediaRecorder(stream);
-            mediaRecorderRef.current = mediaRecorder;
 
-            mediaRecorder.ondataavailable = (event: BlobEvent) => {
-                if (event.data.size > 0) {
-                    recordedChunksRef.current.push(event.data);
-                }
-            };
+            setRecordingSoundFile(audioFile);
+            setValue(fieldName, audioFile as PathValue<T, typeof fieldName>);
+        };
 
-            mediaRecorder.onstop = async () => {
-                const blob = new Blob(recordedChunksRef.current, {
-                    type: "audio/webm",
-                });
-                try {
-                    const base64 = await fileToBase64(blob);
-                    setAudioFile(base64);
-                } catch (error) {
-                    if (error instanceof Error) {
-                        alert(
-                            "Kunde inte konvertera ljudfilen: " + error.message,
-                        );
-                    }
-                    setAudioFile(null);
-                }
-                stream.getTracks().forEach((track) => track.stop());
-            };
-
-            mediaRecorder.start();
-            setIsRecording(true);
-        } catch (err) {
-            if (err instanceof Error) {
-                alert("Kunde inte komma åt mikrofonen: " + err.message);
-                return;
-            }
-            alert("Kunde inte komma åt mikrofonen: " + String(err));
-        }
+        // Stäng av mikrofonen
+        mediaRecorderRef.current.stream.getTracks().forEach((track) => track.stop());
+        setIsRecording(false);
     }
 
-    function stopRecording() {
-        if (mediaRecorderRef.current && isRecording) {
-            mediaRecorderRef.current.stop();
-            setIsRecording(false);
+    function playRecording() {
+        if (!recordingSoundFile) {
+            return;
         }
+        const mediaAudio = new Audio();
+        const file = recordingSoundFile;
+        mediaAudio.src = URL.createObjectURL(file);
+        setAudio(mediaAudio);
+        mediaAudio.play();
     }
 
-    // async function handleAudioInputChange(e: React.ChangeEvent<HTMLInputElement>) {
-    //     const file = e.target.files && e.target.files[0];
-    //     if (file) {
-    //         try {
-    //             const base64 = await fileToBase64(file);
-    //             setAudioFile(base64);
-    //         } catch {
-    //             setAudioFile(null);
-    //         }
-    //     }
-    // }
-
-    function playBase64Audio(base64: string | null) {
-        if (!base64) return;
-        const audio = new window.Audio(base64);
-        audio.play();
+    function pauseRecording() {
+        if (!recordingSoundFile) {
+            return;
+        }
+        if (audio) {
+            audio.pause();
+        }
     }
 
     return {
-        audioFile,
-        isRecording,
         startRecording,
         stopRecording,
-        playBase64Audio,
-        base64ToFileList,
+        isRecording,
+        recordingSoundFile,
+        playRecording,
+        pauseRecording,
     };
 }
+
+export default useAudio2;
